Move DBmysql to Glpi\Database\MySql
