+++ front/backup.php
@@ -169,7 +169,7 @@ function get_content($DB, $table, $from, $limit) {
 
 /**  Get structure of a table
  *
- * @param \Glpi\AbstractDatabase $DB     DB object
+ * @param \Glpi\Database\AbstractDatabase $DB     DB object
  * @param string                 $table  table name
 **/
 function get_def($DB, $table) {
@@ -190,7 +190,7 @@ function get_def($DB, $table) {
 
 /**  Restore a mysql dump
  *
- * @param \Glpi\AbstractDatabase $DB        DB object
+ * @param \Glpi\Database\AbstractDatabase $DB        DB object
  * @param string                 $dumpFile  dump file
  * @param integer                $duree     max delay before refresh
 **/
@@ -308,7 +308,7 @@ function restoreMySqlDump($DB, $dumpFile, $duree) {
 
 /**  Backup a glpi DB
  *
- * @param \Glpi\AbstractDatabase $DB        DB object
+ * @param \Glpi\Database\AbstractDatabase $DB        DB object
  * @param string                 $dumpFile  dump file
  * @param integer                $duree     max delay before refresh
  * @param integer                $rowlimit  rowlimit to backup in one time
