+++ inc/search.class.php
